# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DSCR Loan Automation Platform API
  description: |
    Comprehensive API for the DSCR (Debt Service Coverage Ratio) Loan Automation Platform.

    This platform automates the full lifecycle of DSCR refinance loans from lead intake
    through funding, with native integration to ICE Mortgage Technology Encompass LOS.

    ## Authentication
    All endpoints require Bearer token authentication via OAuth 2.0.

    ## Rate Limiting
    API calls are limited to 1000 requests per minute per client.

    ## Versioning
    The API is versioned via URL path (e.g., /api/v1/).
  version: 1.0.0
  contact:
    name: DSCR Platform Team
    email: support@dscrplatform.com
servers:
  - url: https://api.dscrplatform.com/api/v1
    description: Production
  - url: https://staging-api.dscrplatform.com/api/v1
    description: Staging
...
paths:
...
  /leads:
    post:
      tags: [Leads]
      summary: Create a new lead
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadRequest'
      responses:
        '201':
          description: Lead created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [Leads]
      summary: List leads
      operationId: listLeads
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageToken'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/LeadStatus'
        - name: assignedLoId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of leads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadList'
  /leads/{leadId}:
    get:
      tags: [Leads]
      summary: Get lead by ID
      operationId: getLead
      parameters:
        - $ref: '#/components/parameters/LeadId'
      responses:
        '200':
          description: Lead details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /applications:
    get:
      tags: [Applications]
      summary: List applications
      operationId: listApplications
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageToken'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ApplicationStatus'
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
  /applications/{applicationId}:
    get:
      tags: [Applications]
      summary: Get application by ID
      operationId: getApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /applications/{applicationId}/dscr:
    post:
      tags: [DSCR]
      summary: Calculate DSCR
      operationId: calculateDSCR
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DSCRCalculationRequest'
      responses:
        '200':
          description: DSCR calculation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DSCRCalculationResult'
...
  /applications/{applicationId}/conditions:
    get:
      tags: [Conditions]
      summary: Get conditions for application
      operationId: getConditions
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, CLEARED, WAIVED]
        - name: category
          in: query
          schema:
            type: string
            enum: [PTD, PTC, PTF, POST_CLOSE]
      responses:
        '200':
          description: Conditions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Condition'
...
  /applications/{applicationId}/documents:
    post:
      tags: [Documents]
      summary: Upload document
      operationId: uploadDocument
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'

    get:
      tags: [Documents]
      summary: List documents
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
...
components:
  parameters:
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Application ID
    LeadId:
      name: leadId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Lead ID
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Number of items per page
    PageToken:
      name: pageToken
      in: query
      schema:
        type: string
      description: Pagination token
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    Money:
      type: integer
      description: Amount in cents
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string

    # Lead schemas
    LeadStatus:
      type: string
      enum: [NEW, CONTACTED, QUALIFIED, CONVERTED, LOST, DUPLICATE]
    CreateLeadRequest:
      type: object
      required: [firstName, lastName, email, phone, propertyAddress, loanAmount]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        propertyAddress:
          $ref: '#/components/schemas/Address'
        loanAmount:
          $ref: '#/components/schemas/Money'
        propertyValue:
          $ref: '#/components/schemas/Money'
        monthlyRent:
          $ref: '#/components/schemas/Money'
        creditScoreRange:
          type: string
        source:
          type: string
        tcpaConsent:
          type: boolean
    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/LeadStatus'
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyAddress:
          $ref: '#/components/schemas/Address'
        loanAmount:
          $ref: '#/components/schemas/Money'
        score:
          type: integer
        assignedLoId:
          type: string
        createdAt:
          type: string
          format: date-time
    LeadList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Lead'
        nextPageToken:
          type: string
        totalCount:
          type: integer
...
    ApplicationStatus:
      type: string
      enum: [CREATED, PROCESSING, SUBMITTED, APPROVED, DENIED, FUNDED, CLOSED]
    Application:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        leadId:
          type: string
        borrowerId:
          type: string
        propertyId:
          type: string
        loanAmount:
          $ref: '#/components/schemas/Money'
        interestRate:
          type: number
        loanTerm:
          type: integer
        encompassLoanGuid:
          type: string
        createdAt:
          type: string
          format: date-time
    ApplicationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        nextPageToken:
          type: string
...
    DSCRCalculationRequest:
      type: object
      properties:
        grossMonthlyRent:
          $ref: '#/components/schemas/Money'
        vacancyRate:
          type: number
        managementFee:
          type: number
        propertyTax:
          $ref: '#/components/schemas/Money'
        insurance:
          $ref: '#/components/schemas/Money'
        hoa:
          $ref: '#/components/schemas/Money'
        loanAmount:
          $ref: '#/components/schemas/Money'
        interestRate:
          type: number
        loanTerm:
          type: integer
    DSCRCalculationResult:
      type: object
      properties:
        dscrRatio:
          type: number
        grossRent:
          $ref: '#/components/schemas/Money'
        effectiveRent:
          $ref: '#/components/schemas/Money'
        noiMonthly:
          $ref: '#/components/schemas/Money'
        pitiaMonthly:
          $ref: '#/components/schemas/Money'
        meetsMinimum:
          type: boolean
...
    Condition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        category:
          type: string
          enum: [PTD, PTC, PTF, POST_CLOSE]
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [OPEN, CLEARED, WAIVED]
        priority:
          type: string
          enum: [HIGH, MEDIUM, LOW]
...
    DocumentUploadRequest:
      type: object
      properties:
        file:
          type: string
          format: binary
        documentType:
          type: string
        category:
          type: string
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        documentType:
          type: string
        category:
          type: string
        status:
          type: string
        classificationConfidence:
          type: number
        uploadedAt:
          type: string
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []